Renamed display folder to tmei to reduce confusion.
