+++ include/api.h
@@ -42,7 +42,7 @@
 #define PROS_VERSION_MAJOR 3
 #define PROS_VERSION_MINOR 2
 #define PROS_VERSION_PATCH 1
-#define PROS_VERSION_STRING "3.2.1-dirty.91.e3f5914"
+#define PROS_VERSION_STRING "3.2.1-dirty.94.f52e204"
 
 #define PROS_ERR (INT32_MAX)
 #define PROS_ERR_F (INFINITY)
